Moved types.ts into types dir
